refactor: consume Card and CardLink primitives
