+++ packages/radix/src/index.ts
@@ -4,7 +4,8 @@ export { AspectRatio } from './components/AspectRatio';
 export { Badge } from './components/Badge';
 export { Box } from './components/Box';
 export { Button } from './components/Button';
-export { Card, CardLink } from './components/Card';
+export { Card } from './components/Card';
+export { CardLink } from './components/CardLink';
 export { Checkbox } from './components/Checkbox';
 export { CheckboxButton } from './components/CheckboxButton';
 export { Code } from './components/Code';
